+++ src/cinder/gl/Texture.cpp
@@ -107,16 +107,16 @@ Texture::Texture( const unsigned char *data, int dataFormat, int aWidth, int aHe
 	init( data, 0, dataFormat, GL_UNSIGNED_BYTE, format );
 }
 
-Texture Texture::withCompressedData( const unsigned char *data, int aWidth, int aHeight, int compressedDataSize, Format format )
+Texture Texture::withCompressedData( const unsigned char *data, int aWidth, int aHeight, size_t compressedDataSize, Format format )
 {
-    Texture result;
-    result.mObj = shared_ptr<Obj>( new Obj( aWidth, aHeight ) );
-    if( format.mInternalFormat == -1 )
+	Texture result;
+	result.mObj = shared_ptr<Obj>( new Obj( aWidth, aHeight ) );
+	if( format.mInternalFormat == -1 )
 		format.mInternalFormat = GL_RGBA;
-    result.mObj->mInternalFormat = format.mInternalFormat;
+	result.mObj->mInternalFormat = format.mInternalFormat;
 	result.mObj->mTarget = format.mTarget;
 	result.init( data, compressedDataSize, format );
-    return result;
+	return result;
 }
 
 Texture::Texture( const Surface8u &surface, Format format )
